backend/runner/requirements.txt
torch==1.8.1
torchvision==0.9.1
opencv-python==4.4.0.42
pika==1.2.0
aio-pika==7.2.0
easydict
pyyaml==6.0